Add dependencies for red-black token tree.
Make it easier to switch out the existing splay-tree implementation
with one based on red-black trees.  Motivated by:

    http://lists.racket-lang.org/dev/archive/2012-November/010833.html

According to my analysis, it seems that the pattern of use in
indentation (repeated, sequential search) is bad for the health of a
splay tree.  The typical use of the token-tree matches this worst-case
scenario.